clean up twopence.h and move iostream stuff into a file of its own
